Added environment
And I wasted hours trying to do proper bounding collisions for camera. It ain't working, so back to clamping to a box. I found a MUCH simpler solution: don't try to have complicated scenery within the playable area at all. And I messed with the lighting a bit to make it look way better